Windows update script
I've tested it on Windows 10 and it seems to be doing it's thing :)  It's a .vbs (VBScript) file, since that won't give Windows users a too harsh confrontation by just showing them messageboxes, as opposed to a console window they might get scared from...

The only thing iffy with it is that the URL for more info is embedded in the msgbox, but it isn't clickable nor copy-able.  A solution would be to show an intermediate msgbox asking something like "More info on http:....; would you like to open that page" [y/n], which would open the user's browser at that url.